requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart

# Database
asyncpg>=0.29.0

# Demo parsing & data
awpy>=2.0.0
polars>=1.0.0

# HTTP
httpx>=0.27.0

# Scraping / downloading
playwright>=1.50.0
playwright-stealth>=2.0.0
rarfile>=4.0

# Scheduling
apscheduler>=3.10.0

# Viewer (optional — only needed for pygame viewer)
pygame>=2.6.0
